+++ pysmt/factory.py
@@ -32,6 +32,8 @@ from pysmt.logics import AUTO as AUTO_LOGIC
 from pysmt.logics import most_generic_logic, get_closer_logic
 from pysmt.oracles import get_logic
 
+import pysmt.solvers.solver
+
 DEFAULT_SOLVER_PREFERENCE_LIST = ['msat', 'z3', 'cvc4', 'yices', 'picosat', 'bdd']
 DEFAULT_QELIM_PREFERENCE_LIST = ['z3', 'msat_fm', 'msat_lw']
 DEFAULT_LOGIC = QF_UFLIRA
@@ -43,6 +45,7 @@ class Factory(object):
                  qelim_preference_list=None):
         self.environment = environment
         self._all_solvers = None
+        self._unsat_core_solvers = None
         self._all_qelims = None
         self._generic_solvers = {}
 
@@ -82,7 +85,7 @@ class Factory(object):
                 closer_logic = get_closer_logic(SolverClass.LOGICS, logic)
                 return SolverClass(environment=self.environment,
                                    logic=closer_logic,
-                                   options=None)
+                                   user_options={"generate_models":True})
             else:
                 raise NoSolverAvailableError("Solver %s is not available" % name)
 
@@ -97,10 +100,58 @@ class Factory(object):
             closer_logic = get_closer_logic(SolverClass.LOGICS, logic)
             return SolverClass(environment=self.environment,
                                logic=closer_logic,
-                               options=None)
+                               user_options={"generate_models":True})
         else:
             raise NoSolverAvailableError("No solver is available for logic %s" %\
                                          logic)
+
+
+    def get_unsat_core_solver(self, quantified=False, name=None,
+                              logic=None, unsat_cores_mode="all"):
+        assert quantified is False or logic is None, \
+            "Cannot specify both quantified and logic."
+
+        if quantified is True:
+            logic = self.default_logic.get_quantified_version
+
+        if name is not None:
+            if name in self._unsat_core_solvers:
+                if logic is None:
+                    SolverClass = self._unsat_core_solvers[name]
+                    logic = most_generic_logic(SolverClass.LOGICS)
+                else:
+                    if name in self.all_unsat_core_solvers(logic=logic):
+                        SolverClass = self._unsat_core_solvers[name]
+                    else:
+                        raise NoSolverAvailableError("Solver '%s' does not" \
+                                                     " support logic %s" %
+                                                     (name, logic))
+
+                closer_logic = get_closer_logic(SolverClass.LOGICS, logic)
+                return SolverClass(environment=self.environment,
+                                   logic=closer_logic,
+                                   user_options={"generate_models" : True,
+                                        "unsat_cores_mode" : unsat_cores_mode})
+            else:
+                raise NoSolverAvailableError("Solver %s has no support for " \
+                                             "unsat cores" % name)
+
+        if logic is None:
+            logic = self.default_logic
+
+        solvers = self.all_unsat_core_solvers(logic=logic)
+
+        if solvers is not None and len(solvers) > 0:
+            # Pick the first solver based on preference list
+            SolverClass = self.pick_favorite_solver(solvers)
+            closer_logic = get_closer_logic(SolverClass.LOGICS, logic)
+            return SolverClass(environment=self.environment,
+                               logic=closer_logic,
+                               user_options={"generate_models" : True,
+                                        "unsat_cores_mode" : unsat_cores_mode})
+        else:
+            raise NoSolverAvailableError("No solver supporting unsat cores is" \
+                                         " available for logic %s" % logic)
 
 
 
@@ -134,13 +185,14 @@ class Factory(object):
                                      "eliminator in the preference list")
 
 
-    def add_generic_solver(self, name, args, logics):
+    def add_generic_solver(self, name, args, logics, unsat_core_support=False):
         from pysmt.smtlib.solver import SmtLibSolver
         if name in self._all_solvers:
             raise SolverRedefinitionError("Solver %s already defined" % name)
         self._generic_solvers[name] = (args, logics)
         solver = partial(SmtLibSolver, args)
         solver.LOGICS = logics
+        solver.UNSAT_CORE_SUPPORT = unsat_core_support
         self._all_solvers[name] = solver
 
     def is_generic_solver(self, name):
@@ -151,6 +203,7 @@ class Factory(object):
 
     def _get_available_solvers(self):
         self._all_solvers = {}
+        self._unsat_core_solvers = {}
 
         try:
             from pysmt.solvers.z3 import Z3Solver
@@ -195,6 +248,13 @@ class Factory(object):
         except ImportError:
             pass
 
+
+        for k,s in iteritems(self._all_solvers):
+            try:
+                if s.UNSAT_CORE_SUPPORT:
+                    self._unsat_core_solvers[k] = s
+            except AttributeError:
+                pass
 
 
     def _get_available_qe(self):
@@ -259,6 +319,31 @@ class Factory(object):
 
         return solvers
 
+    def all_unsat_core_solvers(self, logic=None):
+        """
+        Returns a dict <solver_name, solver_class> including all and only
+        the solvers supporting unsat core extraction and directly or
+        indirectly supporting the given logic.  A solver supports a
+        logic if either the given logic is declared in the LOGICS
+        class field or if a logic subsuming the given logic is
+        declared in the LOGICS class field.
+
+        If logic is None, the map will contain all the known solvers
+
+        """
+        res = {}
+        if logic is not None:
+            for s, v in iteritems(self._unsat_core_solvers):
+                for l in v.LOGICS:
+                    if logic <= l:
+                        res[s] = v
+                        break
+            return res
+        else:
+            solvers = self._unsat_core_solvers
+
+        return solvers
+
     def all_qelims(self):
         return self._all_qelims
 
@@ -269,6 +354,14 @@ class Factory(object):
         return self.get_solver(quantified=quantified,
                                name=name,
                                logic=logic)
+
+    def UnsatCoreSolver(self, quantified=False, name=None, logic=None,
+                        unsat_cores_mode="all"):
+        return self.get_unsat_core_solver(quantified=quantified,
+                                          name=name,
+                                          logic=logic,
+                                          unsat_cores_mode=unsat_cores_mode)
+
 
     def QuantifierEliminator(self, name=None):
         return self.get_quantifier_eliminator(name=name)
@@ -291,6 +384,21 @@ class Factory(object):
             if check:
                 retval = solver.get_model()
             return retval
+
+    def get_unsat_core(self, clauses, solver_name=None, logic=None):
+        if logic == AUTO_LOGIC:
+            logic = get_logic(self.environment.formula_manager.And(clauses),
+                              self.environment)
+
+        with self.UnsatCoreSolver(name=solver_name, logic=logic) \
+             as solver:
+            for c in clauses:
+                solver.add_assertion(c)
+            check = solver.solve()
+            if check:
+                return None
+
+            return solver.get_unsat_core()
 
     def is_valid(self, formula, solver_name=None, logic=None):
         if logic == AUTO_LOGIC: