+++ pysmt/solvers/bdd.py
@@ -65,11 +65,11 @@ class LockDdManager(object):
 class BddSolver(Solver):
     LOGICS = [ pysmt.logics.QF_BOOL, pysmt.logics.BOOL ]
 
-    def __init__(self, environment, logic, options=None):
+    def __init__(self, environment, logic, user_options):
         Solver.__init__(self,
                         environment=environment,
                         logic=logic,
-                        options=options)
+                        user_options=user_options)
 
         self.mgr = environment.formula_manager
         self.ddmanager = pycudd.DdManager()